// Angular/Lab-5/src/app/models/student.ts
export interface Student {
  id: number;
  firstName: string;
  lastName: string;
  age: number;
  gender: string;
  email: string;
  phone: string;
  address: string;
  city: string;
  country: string;
  department: string;
  level: number;
  gpa: number;
  enrollmentDate: string;
  isActive: boolean;
}